refactor: use PCM player to avoid conversion
